backend-django/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
requests==2.31.0
gunicorn==21.2.0
